chore(oidc): move account management test to the new `tests` module
