+++ crates/matrix-sdk/src/oidc/mod.rs
@@ -1513,110 +1513,3 @@ fn hash<T: Hash>(x: &T) -> u64 {
     x.hash(&mut hasher);
     hasher.finish()
 }
-
-#[cfg(test)]
-mod tests {
-    use std::str::FromStr;
-
-    use mas_oidc_client::types::{
-        client_credentials::ClientCredentials, registration::ClientMetadata,
-    };
-    use matrix_sdk_base::SessionMeta;
-    use matrix_sdk_test::async_test;
-    use ruma::{
-        api::client::discovery::discover_homeserver::AuthenticationServerInfo, OwnedUserId,
-    };
-    use url::Url;
-
-    use crate::{
-        oidc::{
-            FullSession, OidcAccountManagementAction, RegisteredClientData, SessionTokens,
-            UserSession,
-        },
-        ClientBuilder,
-    };
-
-    #[async_test]
-    async fn test_account_management_url() {
-        let builder =
-            ClientBuilder::new().homeserver_url(Url::parse("https://example.com").unwrap());
-        let client = builder.build().await.unwrap();
-
-        client
-            .restore_session(FullSession {
-                client: RegisteredClientData {
-                    credentials: ClientCredentials::None { client_id: "client_id".to_owned() },
-                    metadata: ClientMetadata {
-                        redirect_uris: Some(vec![Url::parse("https://example.com/login").unwrap()]),
-                        ..Default::default()
-                    }
-                    .validate()
-                    .unwrap(),
-                },
-                user: UserSession {
-                    meta: SessionMeta {
-                        user_id: OwnedUserId::from_str("@user:example.com").unwrap(),
-                        device_id: "device_id".into(),
-                    },
-                    tokens: SessionTokens {
-                        access_token: "access_token".to_owned(),
-                        refresh_token: None,
-                        latest_id_token: None,
-                    },
-                    issuer_info: AuthenticationServerInfo::new(
-                        "https://example.com".to_owned(),
-                        Some("https://example.com/account".to_owned()),
-                    ),
-                },
-            })
-            .await
-            .unwrap();
-
-        assert_eq!(
-            client.oidc().account_management_url(None).unwrap(),
-            Some(Url::parse("https://example.com/account").unwrap())
-        );
-
-        assert_eq!(
-            client
-                .oidc()
-                .account_management_url(Some(OidcAccountManagementAction::Profile))
-                .unwrap(),
-            Some(Url::parse("https://example.com/account?action=profile").unwrap())
-        );
-
-        assert_eq!(
-            client
-                .oidc()
-                .account_management_url(Some(OidcAccountManagementAction::SessionsList))
-                .unwrap(),
-            Some(Url::parse("https://example.com/account?action=sessions_list").unwrap())
-        );
-
-        assert_eq!(
-            client
-                .oidc()
-                .account_management_url(Some(OidcAccountManagementAction::SessionView {
-                    device_id: "my_phone".into()
-                }))
-                .unwrap(),
-            Some(
-                Url::parse("https://example.com/account?action=session_view&device_id=my_phone")
-                    .unwrap()
-            )
-        );
-
-        assert_eq!(
-            client
-                .oidc()
-                .account_management_url(Some(OidcAccountManagementAction::SessionEnd {
-                    device_id: "my_old_phone".into()
-                }))
-                .unwrap(),
-            Some(
-                Url::parse("https://example.com/account?action=session_end&device_id=my_old_phone")
-                    .unwrap()
-            )
-        );
-    }
-}
